+++ includes/admin-options.php
@@ -68,7 +68,7 @@
                 'slug'  => 'advanced_settings',
             );
 
-            if (count(wprss_get_addons()) > 0) {
+            if (count(wprss_get_addons()) > 0 && is_main_site()) {
                 $tabs[] = array(
                     'label' => __( 'Licenses', WPRSS_TEXT_DOMAIN ),
                     'slug'  => 'licenses_settings'
@@ -105,7 +105,17 @@
                 }
                 elseif ( $show_tabs ) {
 
-                    if ( $active_tab === 'licenses_settings' ) {
+                    if ( $active_tab === 'licenses_settings') {
+
+                        if (!is_main_site()) {
+                            printf(
+                                '<p><strong>%s</strong></p>',
+                                __('You do not have access to this page', 'wprss')
+                            );
+
+                            return;
+                        }
+
                         settings_fields( 'wprss_settings_license_keys' );
                         do_settings_sections( 'wprss_settings_license_keys' );
                     }